Add original Py2 version, port to Py3, switch to remote data
This is the second public release v0.6.0, after the first public
release v0.5.0 (named after the fact) was originally published in the
thesis project report.

Everything in the report was produced with v0.5.0, and all the results
and plots should be the same with this v0.6.0.

The difference is that v0.6.0 is ported from Python 2 to Python 3, and
also modified to load the data from a remote repository instead of
locally, to try making it easier for anyone to run the code with the
original data.
An alternative would have been keeping the data in the same Git
repository as the code. But that is in general bad practice, because it
would be untenable, or at least very slow and cumbersome with large
datasets. The caveat about the remote data repository chosen
(data.world) is that you need to create an account there to be able to
access the data. Also, their Python package, for simplifying access to
their REST API, is not updated to more recent versions of Python 3. But
for the time being this seemed like the most simple solution for
enabling remote data, while still hopefully somewhat mainstream. Better
choices could: be directly accessing data.world's REST API, or to choose
another data repository altogether.